array-aho: change payload from void* to int32_t
It is resolved to PyObject in noahong.pyx, the trie only stores the
payload index in the decref list.

It reduces FrozenNode size from 24 bytes to 20, which should reduce our
loaded dataset size by 44M. But I only measure a 10M reduction, I
suspect there is a lot of fragmentation going on.